scraper/requirements.txt
jobspy==1.1.7
supabase==2.4.0
python-dotenv==1.0.0
